Amélioration de la commande de ban, permettant maintenant de ban même un utilisateur non présent sur le serveur où est effectué la commande. + Correctif deleteDPMsgs.js générant des erreurs en cas de messages non existant.

+++ src/utils/answerTmp.js
@@ -3,6 +3,8 @@ const delay = require("./delay");
 
 async function answerTmp(client, message, content, time) {
     const temp_msg = await safeMsgReply(client, message, content);
+    if (temp_msg == null)
+        return;
     await delay(time);
     await temp_msg.delete();
 }
